+++ gulpfile.js
@@ -7,7 +7,7 @@ var
 
 gulp.task('jshint', function () {
 	return gulp
-		.src(['lib/**/*.js', 'test/**/*.js'])
+		.src(['lib/*.js', 'examples/*.js'])
 		.pipe(jshint())
 		.pipe(jshint.reporter('jshint-stylish'));
 });